Allow decimals on items quantities #80

+++ app/Http/Requests/EstimatesRequest.php
@@ -83,11 +83,11 @@ class EstimatesRequest extends FormRequest
                 'required',
             ],
             'items.*.quantity' => [
-                'integer',
+                'numeric',
                 'required',
             ],
             'items.*.price' => [
-                'integer',
+                'numeric',
                 'required',
             ],
         ];
